+++ packages/devextreme/testing/tests/DevExpress.ui.widgets.scheduler/appointment.editing.tests.js
@@ -3,7 +3,7 @@ import $ from 'jquery';
 import { noop } from 'core/utils/common';
 import fx from 'common/core/animation/fx';
 import pointerMock from '../../helpers/pointerMock.js';
-import { hide } from '__internal/ui/tooltip/m_tooltip';
+import { hide } from '@ts/ui/tooltip/tooltip';
 import { DataSource } from 'common/data/data_source/data_source';
 import { Deferred } from 'core/utils/deferred';
 import {
